Resolve a marked defect: duration keywords already tested [402]
Explicitly specified terminal duration as "maturity" in this example
for clarity, even though that keyword was already tested elsewhere.
Aside from that, it's not obvious what the deleted comment meant.
Perhaps its purpose was to suggest adding other duration keywords to
this test, if any were ever added--but none have been.