Treat errors found by InputSequence::realize_vector() uniformly
Where realize_vector() detects errors, an exception (via fatal_error())
is thrown in most cases; however, in one case, a message was formerly
appended to the 'diagnostics' stream...and mark_diagnostic_context() was
called, which logically but fatuously appended:
  Current token 'end of input' at position -1.
because this occurs outside the recursive-descent parser, so there is no
actual parser context. Now, an exception is thrown for any error found
by realize_vector().

This particular test detects out-of-order intervals. The former method
was perhaps appropriate before the recent decision to require intervals
to be specified in order, but now an exception is clearly wanted.

In the unit test, check() has been rewritten to compare expected and
observed diagnostics, whether they're observed by catching an exception
or (regrettably) by calling formatted_diagnostics(). [Use
  git diff --ignore-all-space
to understand this change.]

This commit is a preparatory step that makes ensuing commits simpler and
clearer by confining these members:
  mark_diagnostic_context()
  std::ostringstream diagnostics
to the distinct parser class that aches to be factored out.