Improve [lr]trim unit test and documentation
Added separate tests for ltrim() and rtrim(), and arguably redundant but
cheap tests to show that applying them both to an already-trimmed string
does nothing.

The motivation was not so much to prove that these two functions work
correctly as to provide simple examples for copying and pasting. The
best documentation of a function is a "behold" comment on its unit test:
unlike prose, it's demonstrably correct.