Change group_quote_pdf_generator_wx::footer_ member type and name
This variable contains HTML and not plain text, and so should be of type
wxString and not std::string according to the conventions used in this code.

As it is declared in a class where all the other members are of type
std::string, also rename it to make it even more clear that it is different
from the other fields.