Convert from wxString to std::string explicitly
Don't rely on implicit conversion from wxString to std::string and call
ToStdString(wxConvUTF8) explicitly instead.

The main advantage of doing this is that it avoids silent data loss if
wxString contains any Unicode characters not representable in the current
locale.

Another advantage is that it makes the boundary between GUI and non-GUI code
more clear, albeit at the price of brevity.