Specify expected canonical representations
In most cases, expected and observed canonical representations match.
That part of the tests is still disabled because they don't match in
eleven cases. Of those eleven, most have gaps between intervals, and
the rest all involve ages (or "retirement", which is like an "age");
mathematical_representation() does not yet handle those circumstances.