Make unit tests pastable into GUI
Pasting the text between scissors lines into the GUI, thus:
  File | New | Census
  Census | Paste census
permits exploring differences between class InputSequence and the GUI
sequence editor. Generally, the GUI editor is more strict. Sequences
that it refuses to edit are marked in the "Comments" column. Invalid
sequences are omitted because they cannot even be pasted.