Fix check for switching focus inside InputSequenceEntry
The code checking whether focus remains inside the same InputSequenceEntry
wasn't updated after refactoring of 88f6e9d4aa30b49509fc108dbb269e8718e4cd28
and could return wrong results, e.g. the check failed when focus switched to
the subwindow of wxDataViewCtrl which was the parent of InputSequenceEntry
itself and so the in-place editor wasn't closed in this case, allowing to open
more than one editor at once, which is supposed to be impossible.

Also invert the check direction to make it more natural and use a helper
is_child_of() function to make the parent checks more readable.