Do not paste certain invalid unit tests to census
The first of these two changes removed a line that actually prevented
pasting "census" into a census, because improved error detection in
class InputSequence caused it to throw an exception. The second line
removed did not fulfill its stated purpose, which was to test a logic
error that is not present in production and thus cannot be simulated
by pasting a string into a census.