Prefer brace-or-equal-initializers
sequence() is called by the only ctor, and it cannot initialize
input_stream_ anyway, so there is no reason for it to initialize other
data members that can be initialized in the class declaration.

Formerly, last_input_duration_ was certainly uninitialized sometimes.
That would have been less likely had brace-or-equal-initializers been
available when this code was originally written.