Ensure that parser issues a diagnostic when insanity detected
Formerly, the last statement in span() began
  if(!current_interval_.insane)
which raised the question "else what?". The answer was that presumably
a diagnostic had been issued. Now, if there had been no diagnostic, one
is issued anyway. All diagnostics are treated as errors, so in effect
this ensures that any insanity detected here is trapped.