s12n(c14n)
Simplification of canonicalization. Suppose a thirty-seven-year-old who
plans to retire at age sixty-five wants to take a $1000 one-time-only
withdrawal upon retirement:

  0 retirement; 1000; 0

Formerly, canonicalization transmogrified that into:

  0 retirement; 1000 29; 0

which is obviously correct: 29 = 65 - 37 + 1, "1" being the (implicit)
width of the middle interval. Now, the original string is returned:
still correct, but simpler for end users.

Added a unit test with a comment explaining why
  0 retirement; 1000 #1; 0
is not canonicalized in the same way.