Disable gcc -Wcast-qual warning for wx in the code, not makefile
Don't use imperfect (as indicated by the comment explaining this
technique in the makefile itself as well as the experience described in
http://lists.nongnu.org/archive/html/lmi/2017-03/msg00025.html)
heuristic for determining the files including wxWidgets headers in the
makefile and rely, instead, on the already enforced rule that all source
file using wxWidgets must include pchfile_wx.hpp, even if precompiled
headers are not used.

This allows us to simply disable the warning at the code level, using a
gcc-specific pragma, rather than doing it in an implicit and error-prone
way in the makefiles.