Add [[noreturn]] where 'gcc -Wsuggest-attribute=noreturn' suggests
See:
  http://lists.nongnu.org/archive/html/lmi/2017-03/msg00065.html

Function alarum() correctly does not have the [[noreturn]] attribute
because merely calling it does not throw an exception; an exception is
thrown when the ostream it returns is flushed.

alarum_buf::raise_alert() was given the [[noreturn]] attribute even
though gcc did not suggest it. This is logically required: every
function that can appropriately be assigned to *alarum_alert_function
has attribute [[noreturn]]. An explicit throw-expression was added to
prevent gcc from warning that this function appeared to return.

Incidentally, these warning options
  -Wsuggest-attribute=pure
  -Wsuggest-attribute=const
produced no diagnostics.