Forestall gcc-4.9.2 external attribute warnings, improved
Reverted the 20151221T0138Z change for 'wx_new_test.cpp' only, and
reworked that change in a better way. The unit test explicitly
shared-library attribute, inducing a warning that was suppressed with
a #pragma. Now, it is simply not endowed with any such attribute, so
that no warning arises or requires suppression.