Remove unnecessary definitions of pure virtual functions
These definitions of pure virtual xml_serializable<T> methods were added
in a77e4974b92148155158c23687f58df3424a148e for some unexplained reason,
but don't seem to be necessary because the methods are pure virtual (and
have always been pure, since the initial commit).

As a side effect, this removes the need for using [[noreturn]] attribute
on these methods definitions, which made the program ill-formed as the
attribute was not used with their declarations.