Move [[noreturn]] from definition to declaration
See:
  http://lists.nongnu.org/archive/html/lmi/2017-03/msg00089.html
and the followup discussion. It seems clear that it was wrong to write
the attribute on the function's definition but not on its declaration.
It is less clear whether it can be written on both; this commit follows
the interpretation that it can be written only on a declaration.