Extend bourn_cast to all arithmetic types
Given this major enhancement, it was no longer appropriate to
characterize bourn_cast as a derived work simply because one of the
four conversions it now performs uses the idea Henney expressed in
this single statement on his website:

  return !(arg < 0 && !result_traits::is_signed) &&
         !(arg_traits::is_signed && arg < result_traits::min()) &&
         !(arg > result_traits::max());

although credit is of course still given for that fine work.

Unit tests will follow soon.