Calculate bourn_cast limit directly
Using std::ldexp, calculated 2^N where N is the number of non-sign bits
in the integral type's representation. An integral type's maximum is one
less than this value, and its minimum is either zero or the negative of
this value, decreased in magnitude by one if its representation is not
two's complement.

Replaced the peculiar original limit calculation with this one. Left the
old one in place temporarily, along with code to validate the new limit
against the old.

The new calculation would seem to handle sign-and-magnitude and ones'
complement integers correctly, although that hasn't been tested due to
lack of hardware.

The new calculation also has the virtue that the result of std::ldexp()
is specified (by C99 Annex F) to be an infinity if it overflows, whereas
the behavior of a C++ expression that overflows is undefined according
to C++11 [5.7/6]. Such an infinite result causes all limit tests to
pass, so an extra test for an infinite argument is necessary; this, and
a new test for an unsigned type's zero minimum, changed the exceptions
caught in the unit test, which was adjusted accordingly.