Set msw binary streams in a compiler-independent way
Added '_' prefix as recommended by ms. Apparently mingw has supported
this for a very long time.

Arguably the return code should be checked. However, if it indicates
failure, then the only reasonable thing to do is to print an error
message on one of the output streams that is being reset; but they
must not be written to before they're reset. Furthermore, cgicc
resets them, redundantly (it would appear), which is perhaps an
error itself. Because this code has worked for decades, a tiny and
conservative change seems wise.