Prefer C++11 to C99 transcendentals
Removed workarounds that used, implemented, or approximately emulated
C99 functions
  expm1l()
  log1pl()
in favor of C++11 overloads
  std::expm1(long double)
  std::log1p(long double)
Rooting out all such workarounds utterly was important not only as a
general matter of hygiene, but also because numerous unit tests failed
on x86_86-linux-gnu due to redundant prototypes. For now at least,
the implementations in 'math_functors.hpp' actually call std::expm1l()
and std::log1pl(), so that the long double versions of these functions
are always used even with double arguments, to maintain equivalence to
prior revisions. Although the C++11 standard does not explicitly mention
these 'l'-suffixed functions, it says [26.8/4] that everything in C99
<math.h> is included (subsequent paragraphs describe only additions),
and [17.6.1.2/4] that they're in ns 'std'.