Fix peccable unit tests
This original distich
    // This test would fail for type bool.
    INVOKE_BOOST_TEST( is_strictly_between_extrema(   1), file, line);
was confusing: its test succeeds whatever the containing function
template's type T, even if T is bool (or std::string). The revised test
    is_strictly_between_extrema<T>(1)
does fail for type bool: 'true' is not within the open interval (0,1)
because it equals one of the extrema.

Fixed several other tests that ought to have depended on type T but
formerly did not.