Report unit-test timings as minimum, not mean
When the run time of a function is measured by calling it repeatedly,
the minimum is a better estimate than the mean or mode, as discussed
in the mailing-list message cited in comments as well as
  http://lists.nongnu.org/archive/html/lmi/2017-05/msg00012.html
Both minimum and mean are displayed, but only the minimum is returned.

With a microsecond timer, it didn't make sense to report nanoseconds.

Removed an assertion that a particular unit test takes more than zero
seconds per iteration, because it may take less than one microsecond.