Make wx build directory at a more appropriate time
Extracting an archive creates a source directory, of which the build
directory is a subdirectory. Previously, the build directory was
created before extracting the archive, which was fragile at best. Now
the build directory is created immediately before use.

Incidentally, the source directory is explicitly specified, because
that will become useful soon.