Find utilities on $PATH
A long time ago, lmi was built for msw without cygwin, using dodgy ports
of standard *nix utilities and various nonfree compilers which provided
their own implementations of some utilities such as grep. In that era,
it was necessary to specify the full path to the most suitable version
of each utility; now it is not. For --build=*-pc-cygwin, the cygwin
utilities are all modern, and there is no distinction between /usr/bin/
and /bin/ anyway. For --build=*-unknown-linux-gnu, the Filesystem
Hierarchy Standard doesn't specify the location of every utility: e.g.,
either /bin/grep or /usr/bin/grep is permitted, and hardcoding either
one is a mistake.