Revert "Refactor to accommodate group premium quotes"
This reverts commit c8912c93f392fb92d99ec40f6dc22bb53a2d0a57.

The original purpose was to calculate premiums under assumptions
inconsistent with input--for example, premium with ADB and WP when
those benefits are not elected. That design was just plain weird, and
was thrown away soon after release. Much of the weird code that did
this was eradicated, but this part remained; now it's gone, too.