Refactor to make monthly-deduction discount factor accessible
Formerly, this discount factor was used only for calculating an annuity
rate, so a single function calculated the former and returned the
latter.  Soon the discount factor will be needed for another purpose,
so it deserved a function of its own due to its complexity. The annuity
calculation is so trivial that it is simply inlined.