Add list-bill input and output
The motivation is to use the illustration system to generate list bills
for speculative scenarios. Billing is generally a task for an admin
system, but admin systems generally lack the flexibility required.

Only input and output have been implemented in this commit; calculations
will follow soon. The calculations are complicated because list bills
use a common billing date for contracts that generally don't have a
common anniversary.

Input: Added one new field representing list-bill date. It defaults to
the *nix epoch, chosen on the assumption that no supported inforce
policy was issued on or before that date. Logically, this date cannot
precede InforceAsOfDate: historical list bills cannot be generated
because they'd require historical data that are unavailable. As a
practical matter, though, the input range is not so constrained, and an
illogical date (such as the default) simply results in a billed premium
of zero (which is the most reasonable default outcome, because almost
all end users have zero interest in list bills and will feel least
uncomfortable if "mystery" columns are all zero).

Output: Notionally, one date and one scalar premium have been added.
Following existing practice for similar fields, the date is represented
in two equivalent ways, and the premium is represented as separate
employer and employee portions, and also as their total.