Echo input ListBillDate to output
Added ListBillDate as a group roster column.

The group roster, which is the only report to be used for list bills,
should show the input ListBillDate. It is not easy to show it as a
case invariant, because the roster operates like a *nix filter such as
'sed': it reads one cell at a time as input, processes it, emits output
pertaining to that single cell, and then discards the cell. This means
that it's not simple to ascertain a list bill's validity, which is
suspect if ListBillDate and ErMode are not the same for each cell (and,
thus, for its corresponding output row).

Of course, the group roster could be redesigned to preserve state, as
was done for group quotes. Given that the list bill is a specialized
use case, run only a few times a year, it seems that the cost of that
extra work would exceed the benefit. The tab-delimited output file is
certainly opened in a spreadsheet program, where it is trivial to
validate the consistency of values in a column of arbitrary length in
order to detect input errors.