Do not use value_cast in decimal_root()
This change improved the measured speed of mete_dob_limit() about
sevenfold with i686-w64-mingw32-gcc-4.9.1 by skipping needless work
when writing successive iterands to the default null stream. Reason:
built-in stream inserters exit before performing any formatting when
'badbit' is set, but the (replaced) calls to value_cast<>() could not
be short-circuited in that way.

Setting precision to a large enough value to preserve all information
(i.e., DECIMAL_DIG) is about as good as using value_cast in this case.
Incidentally, 'max_stream_precision.hpp' is no longer needed now that
DECIMAL_DIG is now part of standard C++.

Stream output could be made a little faster by writing a newline
character instead of std::endl. That idea was not used because it
might cause useful information to be lost if the objective function
throws, and because all stream output can be suppressed in the default
null-stream case by making output conditional on 'badbit'.