Resurrect ledger XML IO code as new ledger_evaluator
The code in the new ledger_evaluator.cpp file is a verbatim copy of the
old ledger_xml_io.cpp, however now the computed strings are just stored
inside this class instead of being serialized to XML and a separate
simple class was added to provide access to them.

In the future, this should probably be changed to provide these strings
on demand instead of computing all of them in advance, but for now
minimize the changes compared to the old version of the code.

Update PDF generation code to use ledger_evaluator instead of directly
looking up the values in LedgerInvariant as it did before. This fixes
many problems, included but not limited to duplicating some of the
existing synthetic variables (such as "LmiVersion") and not supporting
other ones (such as "InitTotalSA") at all, and also not using the
correct format for the numeric values in the output. Moreover, this will
make it possible to provide access to vectors and supplemental report
columns too in the upcoming commits.