Simplify an output template
A note at the bottom of every illustration's first page now needs to be
changed in a way that varies by product. This note had formerly been
constructed in the template, where it is difficult to accommodate the
variation now desired. Now it has been moved into the database for each
product, which accommodates such variation easily, allowing the template
to be made much simpler. This change is desirable even for the less
balky template system that will replace XSL.