Use a different graphics format for XSL only
Changed the graphics format from PNG to JPG for XSL only. Reason: a
particular company logo was rendered incorrectly by apache fop-0.20.5,
which has builtin support for JPG but prefers an external library for
PNG--yet does not provide that external library, for licensing reasons:
  http://archive.apache.org/dist/xml/fop/README.html
  | Jimi has been removed for licensing reasons
Work is under way to replace fop with wxPdfDoc, for which PNG is still
used (it works perfectly well, and PNG is preferable to JPG for logos).
Therefore, it would be silly to look into the "jimi" licensing problem
or to upgrade to a version of fop that doesn't use it, and JPG logos are
used as a temporary workaround for XSL only.