Make wx_table_generator even more customizable
Instead of just disabling separators, add a function for using
"condensed style" which also uses smaller row height and avoids the use
of bold font for the headers.

Finally, allow using right alignment for the table columns: this is done
in a separate function because it isn't really related to whether the
style is condensed or not.