Add tabular_details page
This is simple after the refactorings in the previous commits as this
class reuses the logic in the base page_with_tabular_report class.

The only new difficulty is that a hack to support dummy separator
columns is required to reproduce the gap between the guaranteed and
non-guaranteed values in the existing illustrations.