Inline output_and_check_for_page_break() function
This function is used only once, and it doesn't make much sense to have
it in illustration_table_generator class as it doesn't really fit its
role as just a simple wrapper of wx_table_generator defining the
appearance used by the illustration tables.

Doing this also allows to get rid of "page_bottom" argument to this
class ctor, which will make it possible to reuse it when we don't care
about the page bottom and pagination at all.