Get rid of separate wxDC parameter in the code
This DC is always the one associated with pdf_writer_wx, so there is no
need to pass it around when we already pass the writer object itself.

Doing it was thought to be a convenience, but the DC is actually used
very little in the code, so it became a burden instead.