Add numeric summary table to PDF illustrations
Use a custom HTML tag to allow embedding this table in the middle of a
page template.

This solution is more flexible than implicitly appending the table after
the data defined in the external template, as done for the "Tabular
Detail" pages, but doesn't allow having page breaks if the table doesn't
fit on a single page.