Add an assertion, exposing a defect
This assertion fired for an inforce test case with
  <cell version="6">
and a specified amount higher than the specified-amount-load basis.
The defect exposed is in the backward-compatibility code, which uses
the former as a proxy for the latter. This shouldn't matter now in
production if vendor extracts are current and correct, which isn't
necessarily the case. At any rate, the assertion is valuable because
it can detect an implausible value in a current extract.