Use custom <scaled_image> HTML tag for the embedded images
Unfortunately the standard <a> tag can't be used to achieve the scaling
we need as it's limited to integer height and width values, while we
need fractional ones to obtain both the exact image size and aspect
ratio and even the slight difference in the latter results in
catastrophic visual consequences as the image becomes blurred due to
resizing.

Add custom <scaled_image> tag with its "inv_factor" attribute that can
be used to specify exactly the scaling factor needed in HTML, at least
for the values currently being used. And apply these values,
corresponding to the scaling factors of 5/3 and (5/3)^2, to the logos on
the cover page and in the headers of the other pages, respectively.