Reimplement the contract numbers fragment as an external template
This replaces the approach of the previous commit, which generated the
entire fragment in the code, by moving the fragment definition to an
external template and just defining the necessary variables for it in
the code.

This approach is more verbose and arguably less clear, but it keeps the
separation between data and presentation intact.