Add NASD basic illustration page
This is a relatively straightforward reimplementation of the
corresponding part of nasd.xsl.

Implementation reuses the existing page_with_tabular_report class and
the only difference with the other existing tables is that the labels of
the columns of this one are dynamic, i.e. depend themselves on the
illustration values, so need to be generated using HTML interpolator.