Simplify and correct current PDF DC font handling
Only set the smaller font (and although the pen) used for tables once,
when creating the table generator object and avoid changing in it
pdf_writer_wx::output_html() to make it unnecessary to ever set it
again.

This ensures that the same font size is used for the table column widths
determination and actually outputting the text of these columns, which
required explicit use of wxDCFontChanger before.

Also explicitly use 9pt font instead of using wxFont::Smaller() which
resulted in too small font (8pt).