Automatically squeeze table columns to make them fit if necessary
Some tables in the illustrations are so wide that they overflow the page
width if the default generous column margins (~1em) are used, however
globally decreasing the margins for all tables doesn't seem desirable as
they look better with bigger margins.

Solve this problem by making the column margin dynamic and reducing it
automatically if the columns don't fit. This results in optimally
looking tables in all cases and saves time on tweaking the columns
manually to make them fit.