Add "Assumption Detail" page to NASD illustrations
This is relatively straightforward but did require the creation of
"InitAnnLoanDueRate" pseudo-vector in the previous commit.

An alternative approach would be to somehow mark the name of this
variable (e.g. with a leading "=" sign to indicate that is constant
across all rows) and handle such variables specially in the code of
page_with_tabular_report::render().