Factor out common page_with_basic_tabular_report class
NASD and private group placement illustrations share almost the same
basic illustration page, so create a common base class for it which can
be reused for both.

This does create a bit of extra complexity with having to override
get_two_column_header() method, but the gain from not duplicating all
this code in the group private placement case more than compensates for
it.

No real changes so far, this is just a pure refactoring.