Add basic illustration page of group placement illustration
This page implementation is almost completely trivial now thanks to the
preceding page_with_basic_tabular_report refactoring.

Also add footer and header templates for this illustration kind. Notice
that the lower footer part is identical to the one used for NASD
illustrations, but for now keep two copies separate for consistency and
ease of comparison with the existing XSL files.