Use shorter and more consistent names for external templates
Avoid file names longer than 31 character cut off by abbreviating longer
strings in the standard way for all illustrations (e.g. always use
"supp" instead of "supplemental" even for short file names because it
needs to be abbreviated in this way for longer ones) and also use
shorter "reg_d_indiv" prefix for the individual private placement type.

And use "reg_" prefix for the templates and classes used for the regular
illustrations generation for consistency with the other types.