Repaint area covered by SingleChoicePopupMenu after showing it
This is a workaround for a problem of the toolbar of the census view not
repainting after selecting "Print roster to spreadsheet" from the popup
menu shown by one of its toolbar button handlers.

Unfortunately, this is somewhat of a hack, as it involves yielding,
because neither Refresh()-ing nor even Update()-ing the native MSW
toolbar control is sufficient to make it repaint immediately. At least
we can only yield for the UI events, such as wxEVT_PAINT, and so don't
run the risk of reentrancies due to getting some input event, as would
be the case if we just called the global wxYield().

Also notice that, while this fix could be applied just to
census_view.cpp, as the only other current use of SingleChoicePopupMenu
is in docmanager_ex.cpp, where this problem doesn't arise, it seems more
foresighted to do it here, to ensure that it never happens again even if
this class is used elsewhere.