Fall back on generic progress dialog to work around regressions
Class wxProgressDialog has always been specified here, but gcc-4.9.x
caused wxGenericProgressDialog to be used instead. Upgrading to
gcc-6.3.0 caused the native control to be used, resulting in lmi
regressions--see:
  https://trac.wxwidgets.org/ticket/13933
  https://trac.wxwidgets.org/ticket/13185
Reverted the regression by using wxGenericProgressDialog explicitly.
This is only a temporary expedient. Once the problems have been fixed,
the intention is to upgrade wx and revert this commit.