Simplify proprietary-repository workflow
Ensuring that git hooks are symlinked correctly (in all circumstances)
and that git core.filemode is "false" (for msw (cygwin) only) should
prevent the problems that have been observed when an installation that
works well is incompletely migrated to a new machine. Simplified the
normal process by moving manual instructions into a separate section:
they need not be run routinely, but can be consulted if mishaps occur.