Respect minimum when solving for specamt [381]
Solves generally search for a root in [0, 999999999.99]. However,
specified amount usually has a product-specific lower bound greater
than zero. This hasn't seemed to matter in the past, but now it does
with the new NAAR solve: the objective function is ill-conditioned in
that range when solving for specamt to yield a particular target NAAR.

For example, if the minimum specamt is 50000, such a solve might test
iterands such as 0, 1, 10, 100 at the lower end, all of which are
transformed to 50000, producing the same value of the objective
function, and in effect signalling to Brent's algorithm that reducing
the LUB is more profitable than increasing the GLB. Unfortunately, in
realistic test cases, this results in bracketing a discontinuity such
that the contract lapses at the target duration. Respecting the
minimum--in this example, solving in [50000, 999999999.99]--allows
the desired root to be found.