Permit old XSL code to fail
Previously, if the old XSL code failed, the new PDF code wasn't run;
now it is. Motivation: FOP depends on 'java', but the 'java' version
that end users typically use can no longer be installed in a 'wine'
environment--and its installer is not free software, so we can't fix it.