Suppress misleading errors when generating too many PDF pages
The warning about not generating enough pages was also given,
confusingly, when too many of them were generated because the
destructors of the remaining pages were executed as part of stack
unwinding after the failure of the check in next_page() for a previous
page.

Fix this by only giving the warning about "missing extra pages" when the
object is destroyed normally. This fixes the problem described above,
but also makes sense more generally as we don't want to warn about
missing extra pages if the PDF generation failed because of any other
error anyhow, as this is normal and expected in this case, and the
warning just detracts from the real error.