Fix skipping a year in tabular reports in the generated PDF
This fixes a bug introduced by 74872e0aca6f11312f8b65bc49799f8c3a218e28
which forgot to remove the year increment in the outer loop which is not
needed any longer since the year is always incremented, even on reaching
the end of the page, in the inner one.

It also corrects warnings about missing extra pages due to the fact that
this bug resulted in actually generated table being shorter than planned
(which is what 74872e0aca6f11312f8b65bc49799f8c3a218e28 was supposed to
fix in the first place).