Clip variable-width columns in wx_table_generator
This prevents the, potentially arbitrarily long, contents of such
columns from overflowing into the next (and possibly even subsequent)
column(s), making the output completely unreadable.

There is currently no indication that the truncation has occurred, we
might want to show it in some way in the future, e.g. by appending "…"
(U+2026) to the column contents in this case.