Don't rely on wxWindowList being std::list
This is not the case in the default, non-STL, wxWidgets build.

Don't use std::list<> at all (which avoids the need to include and parse
another STL header) and just copy the elements in a temporary vector
instead. Considering the vector size, this extra copy should have no
noticeable effects whatsoever.