Prefer 'a_' prefix to '_' suffix for formal arguments
It is generally better to avoid decorating arguments at all. In cases
where decoration improves clarity, e.g.:
  foo::foo(int max) : max(max) ...
an 'a_' prefix may reluctantly be used. Never use a '_' suffix to
distinguish arguments: that is reserved for member variables.

As long as the '_' suffix is used for all data members and never used
for any other purpose, the example above is naturally written thus:
  foo::foo(int max) : max_(max) ...
and one would never think of writing any 'a_' prefix.