Fix and simplify simplify_floating_point()
Don't dereference an invalid iterator for strings consisting only of
zeroes in simplify_floating_point(): the reverse iterator was
incremented after reaching the beginning of the string, making it
invalid, and then used in std::string ctor.

Fix this by avoiding the confusing goto inside switch statement and
using a straightforward loop instead.