Sarculate and weed out redundant repetition
There actually is a serious reason for this simplification: it appears
that some extra cleanup step might be wanted for input files provided
by a certain third-party vendor, and the sanity-assertion function may
suggest where to execute that step. Calling that function redundantly
is of no practical consequence today because it is so lightweight, but
repeating a costly extra step would impair responsiveness.