Suppress a harmless gcc-7 warning
Bourn-casting a negative floating value to bool elicits an exception:
  if(!to_traits::is_signed && from < 0) throw...
In that case, the unreachable next statement compares such a value to
the lower limit for type bool, calculated in a way that depends on
whether bool is represented as two's complement, as determined by a
technique that involves the bitwise negation of the target type (bool
in this case). IOW, all negatives having already been rejected, the
question at hand is whether the lowest acceptable negative is -1 or -2.

It is nevertheless dubious to form the bitwise negation of a bool,
even where it's harmless. That will be easier to avoid once gcc-7 is
accepted for production.