Refine instructions for installing native msw software in chroot
Indicated a convenient moment for installing native msw software.

Removed instructions for installing msw-native java: they don't work
since debian-9 because the installer most commonly used for msw says:
  Install Failed
  This machine does not have the recommended amount of memory to
  install this product.
on a machine with 64 GiB of RAM, and that cannot be fixed because,
while java is said to be free, this installer is not; and once FOP
has been replaced, end users won't need java to run lmi anyway.