Improve concinnity
Write a comment explaining the condition for...
  ...the code that follows  an #else  directive
  ...the code that precedes an #endif directive
thus:

  #if condition
  #endif // condition

  #if condition
  #else  // !condition
  #endif // !condition

Prefer '#if defined FOO' to '#ifdef FOO' because it makes the condition
easy to copy to corresponding '#else' and '#endif' directives. Likewise
prefer '#if !defined FOO' to '#ifndef FOO'. This would be misleading:

  #ifdef FOO
  #endif // FOO

because the code above '#endif' depends not on whether FOO evaluates to
nonzero, but whether it is defined. Exception: write header guards for
'bar.hpp' thus:

  #ifndef timer_hpp
  #define timer_hpp
  ...
  #endif // timer_hpp

to emphasize their distinctive purpose.