Further document a MinGW-w64 gcc-7.2.0 anomaly
floor() and trunc() yield different results for a particular floating-
point argument slightly different from positive three. To reproduce:
  File | New | Illustration
  "Face" tab: append "000" to "Specified amount"
  "Payments" tab: append "000" to "Individual payment"
  OK
  File | Print to PDF
Values are scaled by 10^3N when they would be too enormous to fit in
a column with print mask "999,999,999"; in this case, 3N is about 3.
(It's calculated as -log10(0.001), which isn't exactly int(3).)

Observed output produced by this stream insertion:
    << "\n  " << (power)             << " logarithm"
    << "\n  " << (std::floor(power)) << " floor()"
    << "\n  " << (std::trunc(power)) << " trunc()"
is
  3 logarithm
  2 floor()
  3 trunc()
but floor() and trunc() should give the same result.