Work around a MinGW-w64 gcc-7.2.0 anomaly
Reverted commits 1c1bafa402 and d584c2a02d, and suppressed a failing
assertion.

The d584c2a02d commit message said:
| For integral N, log10(10^N) is not an exact integer.
and that is true; but the problem here is that N was negative, so
10^N was not an exact integer as might be expected for positive N.
Thus, while it might be hoped that log10(1000) would return exactly 3,
it is less certain that log10(0.001) exactly equals -3.

Now that the nature of the problem is understood, the elaborate
diagnostic message recently added is no longer wanted. The root cause,
of course, is that 0.001 is reliably producible from 3, but 3 is not
reliably recoverable from 0.001 .